// src/modules/discussion/discussion.enum.ts
export enum DiscussionType {
  PROJECT_TASK = 'project_task',
  PROJECT_TASK_PROPOSAL = 'project-task-proposal',
  USER_MOVEMENT = 'user_movement',
}

export enum DiscussionTagType {
  ALL = 'all',
  USER = 'user',
}

export enum DiscussionReactionType {
  LIKE = 'like',
  LOVE = 'love',
  HAHA = 'haha',
  WOW = 'wow',
  SAD = 'sad',
  ANGRY = 'angry',
}